Add missing unittests
These unittest were in the internal repo but not in the public one. It
is uncertain if they were not added to the public repo because of some
test failing due to different test environments or if they were added
later to the internal repo and they were lost in the sync process.